data/requirements.txt
duckdb>=1.1.0
pyarrow>=17.0.0
python-chess>=1.999
